open-lessons/django-and-drf-17.04/pyproject.toml
[project]
name = "django-and-drf-17-04"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2",
    "djangorestframework>=3.16.0",
]
